+++ src/main/java/roomescape/global/auth/AuthUser.java
@@ -1,0 +1,6 @@
+package roomescape.global.auth;
+
+public record AuthUser(
+        Long id
+) {
+}
