--- src/main/java/roomescape/global/config/AuthenticationPrincipal.java
@@ -1,12 +1,0 @@
-package roomescape.global.config;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticationPrincipal {
-    boolean required() default true;
-}
